Refactors into proper project with cli tools
