+++ bin/ig65m
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python3 -m ig65m.cli "$@"
